==> notebooks/05_create_profit_chart.ipynb <==
#Last edited/used: 5/11/2023

import pandas as pd
import pandas_ta as ta
import os
import plotly.graph_objects as go
import pickle
import datetime

# Set horizon
horizon = 'SCALP'
# horizon = 'SWING'
# horizon = 'POSITION'

PATH = os.path.dirname(os.getcwd())
data_path = os.path.join(PATH, 'data')

#SCALP
# W30_H2_DF30_V67_SCALP_04_12_2023_PROFIT_MAP.pkl
#SWING
# W30_H4_DF1W_V00_SWING_04_10_2023_PROFIT_MAP.pkl
#POSITION 
# W30_H4_DF1W_V00_POSITION_04_10_2023_PROFIT_MAP.pkl


if horizon == 'SCALP':
    file_name = 'W30_H2_DF30_V67_SCALP_04_12_2023_PROFIT_MAP.pkl'
elif horizon == 'SWING':
    file_name = 'W30_H4_DF1W_V00_SWING_04_10_2023_PROFIT_MAP.pkl'
elif horizon == 'POSITION':
    file_name = 'W30_H4_DF1W_V00_POSITION_04_10_2023_PROFIT_MAP.pkl'


file_path = os.path.join(data_path, file_name)
# Open and Read file
with open(os.path.join(data_path, file_name), 'rb') as f:
    loaded_dict = pickle.load(f)

long = pd.DataFrame(loaded_dict["LONG"])
short = pd.DataFrame(loaded_dict["SHORT"])
frames = [long, short]

profit_chart = pd.concat(frames, axis=0, ignore_index=True)
profit_chart = profit_chart.sort_values(by=['PredictionDate']).reset_index(drop=True)

profit_chart.head()

len(profit_chart)

profit_chart.to_csv('../data/{}_profit_chart.csv'.format(horizon))
